// src/app/models/notification-type.enum.ts
// src/app/models/notification-type.enum.ts
export enum NotificationType {
  // MATCH
  MATCH_CREATED = 'MATCH_CREATED',

  // MESSAGE
  MESSAGE_RECEIVED = 'MESSAGE_RECEIVED',

  // DELIVERABLE
  DELIVERABLE_SENT = 'DELIVERABLE_SENT',
  DELIVERABLE_VALIDATED = 'DELIVERABLE_VALIDATED',
  DELIVERABLE_REJECTED = 'DELIVERABLE_REJECTED',

  // PAYMENT / TRANCHE
  TRANCHE_LINK_READY = 'TRANCHE_LINK_READY',
  TRANCHE_PAID = 'TRANCHE_PAID',
  TRANCHE_VALIDATED = 'TRANCHE_VALIDATED',
  TRANCHE_ERROR = 'TRANCHE_ERROR',

  // WITHDRAWAL
  WITHDRAWAL_PAID = 'WITHDRAWAL_PAID',
  WITHDRAWAL_ERROR = 'WITHDRAWAL_ERROR',

  // MISSION
  MISSION_STATUS_CHANGED = 'MISSION_STATUS_CHANGED',
  MISSION_DEADLINE_SOON = 'MISSION_DEADLINE_SOON'
}